+++ OnlineExamSystem/src/main/java/com/exam/dao/QuestionDao.java
@@ -9,7 +9,7 @@ import com.exam.entity.exam.Questions;
 import com.exam.entity.exam.Quiz;
 
 @Repository
-public interface QuestionDao extends JpaRepository<Questions, Long>{
+public interface QuestionDao extends JpaRepository<Questions, Integer>{
 
 	Set<Questions> findByQuiz(Quiz quiz);//the findBy should be as it is but the quiz variable should match the variable declared at Questions pojo as a camelcase
 	//custom finder method
