+++ OnlineExamSystem/src/main/java/com/exam/service/CategoryService.java
@@ -9,7 +9,7 @@ public interface CategoryService {
 	public Category addCategory(Category category);
 	public Category updateCategory(Category category);
 	public Set<Category> getCategories();
-	public Category getCategory(Long categoryId);
-	public void deleteCategory(Long categoryId);
+	public Category getCategory(int categoryId);
+	public void deleteCategory(int categoryId);
 
 }
