+++ OnlineExamSystem/src/main/java/com/exam/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService {
 	//get user by username
 	public User getUser(String username);
 	//delete user by id
-	public void deleteUser(Long userId);
+	public void deleteUser(int userId);
 	
 
 }
